+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+ - Fixed a mistake that was preventing Occupancy Sensors from being instantiated
+
 ## [6.4.0] - 2019-07-25
 ### Changed
  - Split CsaPws10sHubEnetAdapter into two master - slave adapters and changed interface implementations to match